// src/core/models/enums/app-color.ts
export enum AppColor {
    blue = '#157DF1',
    darkBlue = '#0057B9',
    lightBlue = '#4995F1',
    almostWhiteBlue = '#ABD2FF',
    white = '#FFFFFF',
    black = '#333333',
    grey = '#808080',
    red = '#F0425F',
    green = '#3DCA6A',
    lightGrey = '#EDF3FA',
    contrastGrey = '#E8E9EB',
    searchGrey = '#F2F3F5',

    // #region Table colors
    tableHead = '#E5EBF2',
    tableRowDark = '#F8F9FA',
    tableItemRed = '#E63964',
    tableInactive = '#C4C4C4',
    // #endregion
}

export enum GraphicColor {
    darkestBlue = '#3C4BF2',
    darkBlue = '#157DF1',
    blue = '#509DF4',
    whiteBlue = '#C5DFFB',
    lightBlue = '#8ABEF8',
    red = '#E53964',
    pink = '#F6BCCA',
    green = '#3DCA6A',
    lightGreen = '#D0F2DB',
    orange = '#FB9631',
    orangeLight = '#FEE7D0',
}